perks are finished, some perks need checking but descriptions and most perks are in
