+++ SluaghEngine/includes/Gameplay/perkConditionEnum.h
@@ -1,0 +1,22 @@
+#pragma once
+namespace SE
+{
+	namespace Gameplay
+	{
+		enum class perkConditions
+		{
+			NONE = 0,
+			PHYSICAL_TAKEN = 1,
+			WATER_TAKEN = 2,
+			NATURE_TAKEN = 3,
+			MAGIC_TAKEN = 4,
+			FIRE_TAKEN = 5,
+
+			PHYSICAL_GIVEN = 6,
+			WATER_GIVEN = 7,
+			NATURE_GIVEN = 8,
+			MAGIC_GIVEN = 9,
+			FIRE_GIVEN = 10
+		};
+	}
+}
